al_btree: Add option for inplace operations
To cut down on the data structure size add the explicit option for
inplace operations using the array list pointer as the list itself. This
allows for cutting down the data we need to store in the data structure
which will optimize space for the places where we use these structures a
lot.

Update poolalloc to use this implementation, since it was already using
it in this static way.

Signed-off-by: Luke Nowakowski-Krijger <lukenowakow@proton.me>