common/al_btree: Rewrite to 64bit list by default
Make the data structure simpler by making it a single 64 bit list by
default, which gives us a static binary tree array list and makes the
adds/deletes bounded and resonable while using as little space as
possible.

This lets the other places that use these lists as compact as possible
since they will be using a lot of them.

Also simplify some of the exported functions/defines to make it more
readable and simpler.

Signed-off-by: Luke Nowakowski-Krijger <lukenowakow@proton.me>