kernel/mm: Implement buddy allocation of physical pages
Implement the buddy allocation of physical pages and define the general
memory areas and data structures needed to keep track of physical
frames.

There is still work to be done in how buddy nodes are searched for and
how the buddy tree is traversed, in particular use of the mm_buddy_dir_t
that shows which "direction" of the tree and which buddy to select can
be removed in favor of just parsing the addresses in a smarter way.

There is also a lot of bloat in the #defines in the header that can and
will be trimmed down.

Signed-off-by: Luke Nowakowski-Krijger <lukenowakow@proton.me>