kernel/kalloc: Reimplemnt kalloc to use buddy allocation
Reimplement kalloc to use the new buddy allocation scheme for pages as
well as using kalloc_cache/slab imeplementation for a more efficent and
faster allocing of smaller sized objects.

The previous implementation using the arraylist_btree struct was not
good enough since it used the same data for allocing small objects and
for pages, and since we need page allocing to be fast and to have a more
central data structures for pages available, I went in favor of Linux's
approach to memory allocation structs which I found to be more efficent.

Signed-off-by: Luke Nowakowski-Krijger <lukenowakow@proton.me>