Implement proper buddy allocation support in kernel/kalloc_page.c
Reimplment buddy allocation using a statically defined buddy array that
maps 1 buddy to every two pages. This gives us a static lookup of a page
allocation based on its page index, and lets us avoid using a bitmap and
needing to check free_lists in a linear manner. The memory footmap is
rather large, but we can trim it down if we implement smaller version of
singly linked lists only used for linking, instead of the doubly linked
list with data elements that take up 60% of the memory footprint and is
unused.

This also lets us circumvent any array index like logic, and have the
logic more around page indexs and memorders. This is very much how Linux
does it and I liked the implementation because it was very
straightforward.

The rest of the logic for allocing and freeing is pretty
straightforward, with most of the more complicated logic around
allocating at a specific address, which I wanted to implement for
completness and for reservering specific pages later on for memory
mapped segments and specific memory ranges I would want to reserve.

There is a lot of code around allocating and freeing pages of single
pages, which we needed to multiplex single pages from the buddy which
covers 2 pages.

All in all a better and cleaner implementation.

Signed-off-by: Luke Nowakowski-Krijger <lukenowakow@proton.me>