Move kernel code to higher addresses
Implement early mmu initialization in early_init and early_mm since
these will be in the early text section that we have access to at boot.
Initiliaze the mmu and provide and early environment for allocating
memory.

Change the linker addresses to our higher kernel addresses and create
new sections for early text and data.

Signed-off-by: Luke Nowakowski-Krijger <lukenowakow@proton.me>