common/linked_list: Implement multi type linked lists
Implement multi typed linked lists so that we can conserve on memory
where we need to while having most of the code centralized in the same
library.

Especially for the memory manager buddies that each had a linked list
node that had last and data fields that were not being used in the
previous implementation, we can now cut down on two thirds of the
footprint by implementing more leaner linked lists types and using the
doubly linked list types when we need them.

Signed-off-by: Luke Nowakowski-Krijger <lukenowakow@proton.me>