kernel/early_mm: Add physical memory map and properly reserve pages
Add a physical memory map, even though it is pretty hardcoded, that we
can use for getting memory sizes and attributes for the mmu and mm
setup.

Export the early mm heap so that we can properly reserve pages up to our
early memory space, which now means that all the free space in the mm
buddy allocator is actually free memory that can be allocated.

Signed-off-by: Luke Nowakowski-Krijger <lukenowakow@proton.me>