kernel/early_init: Fix early mem map null entry
When the physical address mapping passed 0x40000000 we would not be
correctly handling the accessing of the second block address, as it
would return null on the level0 block entry and never be correctly set.

Also fix a bug that if a memory map entry range is greater than or equal
to the block size that it will still be mapped.

Signed-off-by: Luke Nowakowski-Krijger <lukenowakow@proton.me>