kernel/early_init: Correct MMIO range and map QA7 registers
Correctly calculate the MMIO range as it was incorectly calculate
before. In reality we could just pull these values from the DTB but
these hand calculations are okay for now.

Also map the Quad cortex-A7 mmio registers as an entry to have their
memory mapped range added to the page table.

Signed-off-by: Luke Nowakowski-Krijger <lukenowakow@proton.me>